requirements.txt
hachoir==3.2.0
numpy==1.24.2
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
pytesseract==0.3.10
